Extract isPrefixPresent out to ArgumentMultimap
