+++ src/main/java/seedu/taassist/logic/parser/ArgumentMultimap.java
@@ -1,10 +1,17 @@
 package seedu.taassist.logic.parser;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
+
+import static java.util.Objects.requireNonNull;
+import seedu.taassist.logic.parser.exceptions.ParseException;
+import seedu.taassist.model.moduleclass.ModuleClass;
 
 /**
  * Stores mapping of prefixes to their respective arguments.
@@ -56,5 +63,13 @@ public class ArgumentMultimap {
      */
     public String getPreamble() {
         return getValue(new Prefix("")).orElse("");
+    }
+
+    /**
+     * Returns true if the {@code prefix} does not contain empty {@code Optional} values in the given
+     * {@code ArgumentMultimap}.
+     */
+    public static boolean isPrefixPresent(ArgumentMultimap argumentMultimap, Prefix prefix) {
+        return argumentMultimap.getValue(prefix).isPresent();
     }
 }
